New system prompt style
This commit implements a new system prompt style, where the pre-task engagement is considered a task and treated like that.
It also format a few prompts to fit into the refactoring we are going to implement on the AI Tutor, with the aim of including only relevant bits when needed and reducing the total number of imput tokens for each request.